chore: move loadConfig to node_loader
