+++ lib/node_loader.js
@@ -18,8 +18,36 @@ util.clientSideMonitoring = {
   Publisher: require('./publisher').Publisher,
   configProvider: require('./publisher/configuration'),
 };
+var iniLoader = require('./shared-ini').iniLoader;
 util.iniLoader = require('./shared-ini').iniLoader;
 util.getSystemErrorName = require('util').getSystemErrorName;
+
+util.loadConfig = function(options) {
+  var envValue = options.environmentVariableSelector(process.env);
+  if (envValue !== undefined) {
+    return envValue;
+  }
+
+  var configFile = {};
+  try {
+    configFile = iniLoader ? iniLoader.loadFrom({
+      isConfig: true,
+      filename: process.env[util.sharedConfigFileEnv]
+    }) : {};
+  } catch (e) {}
+  var sharedFileConfig = configFile[
+    process.env.AWS_PROFILE || util.defaultProfile
+  ] || {};
+  var configValue = options.configFileSelector(sharedFileConfig);
+  if (configValue !== undefined) {
+    return configValue;
+  }
+
+  if (typeof options.default === 'function') {
+    return options.default();
+  }
+  return options.default;
+};
 
 var AWS;
