+++ lib/util.js
@@ -1000,33 +1000,6 @@ var util = {
     }
   },
 
-  loadConfig: function loadConfig(options) {
-    var envValue = options.environmentVariableSelector(process.env);
-    if (envValue !== undefined) {
-      return envValue;
-    }
-
-    var configFile = {};
-    try {
-      configFile = AWS.util.iniLoader ? AWS.util.iniLoader.loadFrom({
-        isConfig: true,
-        filename: process.env[AWS.util.sharedConfigFileEnv]
-      }) : {};
-    } catch (e) {}
-    var sharedFileConfig = configFile[
-      process.env.AWS_PROFILE || AWS.util.defaultProfile
-    ] || {};
-    var configValue = options.configFileSelector(sharedFileConfig);
-    if (configValue !== undefined) {
-      return configValue;
-    }
-
-    if (typeof options.default === 'function') {
-      return options.default();
-    }
-    return options.default;
-  },
-
   /**
    * @api private
    */
